added heatmaps notebook
